+++ packages/venia-ui/lib/components/Navigation/navHeader.js
@@ -10,35 +10,35 @@ import { mergeClasses } from '../../classify';
 import Icon from '../Icon';
 import Trigger from '../Trigger';
 import defaultClasses from './navHeader.css';
-
-const titles = {
-    CREATE_ACCOUNT: 'Create Account',
-    FORGOT_PASSWORD: 'Forgot Password',
-    MY_ACCOUNT: 'My Account',
-    SIGN_IN: 'Sign In',
-    MENU: 'Main Menu'
-};
+import { useNavigationHeader } from '@magento/peregrine/lib/talons/Navigation/useNavigationHeader';
 
 const NavHeader = props => {
     const { isTopLevel, onBack, onClose, view } = props;
-    const classes = mergeClasses(defaultClasses, props.classes);
 
-    const title = titles[view] || titles.MENU;
-    const backIcon = isTopLevel && view === 'MENU' ? MenuIcon : ArrowLeftIcon;
+    const talonProps = useNavigationHeader({
+        isTopLevel,
+        onBack,
+        onClose,
+        view
+    });
 
-    const backButton = !(isTopLevel && view === 'MENU') ? (
-        <Trigger key="backButton" action={onBack}>
+    const { handleClose, handleBack, isTopLevelMenu, title } = talonProps;
+
+    const backIcon = isTopLevelMenu ? MenuIcon : ArrowLeftIcon;
+    const backButton = !isTopLevelMenu ? (
+        <Trigger key="backButton" action={handleBack}>
             <Icon src={backIcon} />
         </Trigger>
     ) : null;
 
+    const classes = mergeClasses(defaultClasses, props.classes);
     return (
         <Fragment>
             {backButton}
             <h2 key="title" className={classes.title}>
                 <span>{title}</span>
             </h2>
-            <Trigger key="closeButton" action={onClose}>
+            <Trigger key="closeButton" action={handleClose}>
                 <Icon src={CloseIcon} />
             </Trigger>
         </Fragment>
